[Front] refactor(skills): move makeNewSkillEditorsGroup to SkillResource
